requirements.txt
streamlit==1.28.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
tabulate==0.9.0

